Backport code that supports Elm 0.18 (#18)
This follows the pattern used in elm-css to maintain support for Elm 0.18 along
with the new support for Elm 0.19.

The new version that supports Elm 0.19 will be 3.0.2. The goal is that version 3.0.2 makes
the Elm migration process easier for users of this package, since it will work in both 0.18 and 0.19.
My intent is for that to be the final version that supports Elm 0.18,
unless there are additional required changes in the near future.

Ideally, the next desired update to this package will not occur until after Elm 0.19
has wider adoption, and therefore dropping support for Elm 0.18 will not be a burden.
That seems reasonable since this package has been stable for the last year.